+++ src/lightly_train/_models/dinov2_vit/dinov2_vit_package.py
@@ -59,10 +59,15 @@ class DINOv2ViTPackage(Package):
         Get a DINOv2 ViT model by name. Here the student version is build.
         """
         # Replace "_" with "-" for backwards compatibility.
-        # - "vitb14_pretain" -> "vitb14-pretrain"
-        # - "_vittest14_pretrain" -> "_vittest14-pretrain"
+        # - "vitb14_pretained" -> "vitb14-pretrained"
+        # - "_vittest14_pretrained" -> "_vittest14-pretrained"
         # We keep leading underscores for private test models.
         model_name = model_name[0] + model_name[1:].replace("_", "-")
+
+        # Replace "-pretrain" with "-pretrained" suffix for backwards compatibility.
+        if model_name.endswith("-pretrain"):
+            model_name = model_name[:-len("-pretrain")] + "-pretrained"
+
         if model_name not in VIT_MODELS:
             raise ValueError(
                 f"Unknown model: {model_name} available models are: {cls.list_model_names()}"
@@ -104,7 +109,7 @@ class DINOv2ViTPackage(Package):
         model = model_builder(**kwargs)
 
         # Load the pretrained model if required
-        if model_name.endswith("-pretrain"):
+        if model_name.endswith("-pretrained"):
             cache_dir = get_cache_dir()
             checkpoint_dir = cache_dir / "weights"
             model = load_weights(