+++ src/lightly_train/_models/dinov2_vit/dinov2_vit_src/configs/__init__.py
@@ -34,19 +34,19 @@ MODELS = {
         "url": "",
         "config": "train/vitg14",
     },
-    "vits14-pretrain": {
+    "vits14-pretrained": {
         "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vits14/dinov2_vits14_pretrain.pth",
         "config": "eval/vits14_pretrain",
     },
-    "vitb14-pretrain": {
+    "vitb14-pretrained": {
         "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vitb14/dinov2_vitb14_pretrain.pth",
         "config": "eval/vitb14_pretrain",
     },
-    "vitl14-pretrain": {
+    "vitl14-pretrained": {
         "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vitl14/dinov2_vitl14_pretrain.pth",
         "config": "eval/vitl14_pretrain",
     },
-    "vitg14-pretrain": {
+    "vitg14-pretrained": {
         "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vitg14/dinov2_vitg14_pretrain.pth",
         "config": "eval/vitg14_pretrain",
     },