+++ CHANGELOG.md
@@ -14,7 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Patch embedding in ViT to support image resolution that is not a multiple of the patch size.
 - Augmentations plotting function to support odd-sized images.
 - DistillationV2 to work with image resolution that is not a multiple of the patch size.
-- DINOv2 model names changed from `dinov2_vit/<model_name>` to `dinov2/<model_name>`.
+- DINOv2 model names changed from `dinov2_vit/vits14_pretrain` to `dinov2/vits14-pretrained`
+  to make the naming more consistent. Old names are still supported but will be removed
+  in a future release.
 
 ### Deprecated